image-search/requirements.txt
ucall
uform
usearch

numpy
pillow